from dev/src * Houdini Engine/HAPI: changed all of the time related variables to doubles to match time being specified as an fpreal within Houdini (apart from float builds). This fixes issues with lack of precision showing up in frame numbers (instead of 101, getting frame 100.999999), which was causing issues for customers. * Houdini Engine for 3ds Max: updated the HAPI bindings * Houdini Engine for 3ds Max: swept the plugin to change time variables from float to double in order to reflect the change in HAPI * Houdini Engine for Unreal: updated HAPI bindings * HoudiniEngineSample: updated HAPI bindings * Houdini Engine for Maya: updated HAPI bindings * Houdini Engine for Maya: updated the plugin to use double to represent time instead of float, in order to reflect the changes in HAPI * Houdini Engine for Unity: updated the GetTime and SetTime methods to use doubles instead of floats to reflect the change in HAPI

+++ Plugins/HoudiniEngineUnity/Scripts/Sessions/HEU_SessionBase.cs
@@ -547,12 +547,12 @@ namespace HoudiniEngineUnity
 
         // TIME -----------------------------------------------------------------------------------------------------
 
-        public virtual float GetTime()
+        public virtual double GetTime()
         {
             return 0;
         }
 
-        public virtual bool SetTime(float time)
+        public virtual bool SetTime(double time)
         {
             return false;
         }
